restructure files for readability
